Test for equipped axe when approaching the tree ...
It is natural to equip an axe to chop a tree, this tests for "iron axe"
or "fine iron axe" are equipped, in addition to them being held in
inventory.

It would confuse players if they approach the tree and be told you need
an axe, when you are wielding it.